+++ src/directive/angular-tree-dnd-compile.js
@@ -32,4 +32,32 @@ angular.module('ntt.TreeDnD')
                 }
             };
         }]
+)
+    .directive(
+    'compileReplace', [
+        '$compile', function ($compile) {
+            return {
+                restrict: 'A',
+                link:     function (scope, element, attrs) {
+                    scope.$watch(
+                        attrs.compileReplace, function (new_val) {
+                            if (new_val) {
+                                /*
+                                 * Compile creates a linking function
+                                 * that can be used with any scope.
+                                 */
+                                var link = $compile(new_val);
+                                /*
+                                 * Executing the linking function
+                                 * creates a new element.
+                                 */
+                                var new_elem = link(scope);
+
+                                element.replaceWith(new_elem)
+                            }
+                        }
+                    );
+                }
+            };
+        }]
 );