+++ src/directive/angular-tree-dnd-node.js
@@ -4,23 +4,11 @@ angular.module('ntt.TreeDnD')
         return {
             restrict: 'A',
             replace:  true,
-            link:       function (scope, element, attrs) {
+            link:     function (scope, element, attrs) {
+                var _enabledDragDrop = (typeof scope.dragEnabled === 'boolean' || typeof scope.dropEnabled === 'boolean');
                 scope.$modelValue = null;
-                scope.$element = element;
-                scope.$type = 'TreeDnDNode';
                 scope.$icon_class = '';
                 scope.$node_class = '';
-                scope.getScopeNode = function () {
-                    return scope;
-                };
-
-                scope.getData = function () {
-                    return scope.$modelValue;
-                };
-
-                scope.getElementChilds = function () {
-                    return angular.element(element[0].querySelector('[tree-dnd-nodes]'));
-                };
 
                 if (scope.$class.node) {
                     element.addClass(scope.$class.node);
@@ -29,11 +17,31 @@ angular.module('ntt.TreeDnD')
 
                 scope.$watch(
                     attrs.treeDndNode, function (newValue, oldValue, scope) {
-                        scope.setScope(scope, newValue);
+                        if (_enabledDragDrop) {
+                            scope.setScope(scope, newValue);
+                        }
                         scope.$modelValue = newValue;
                         scope.$icon_class = scope.$class.icon[newValue.__icon__];
                     }, true
                 );
+
+                if (_enabledDragDrop) {
+
+                    scope.$element = element;
+                    scope.$type = 'TreeDnDNode';
+
+                    scope.getScopeNode = function () {
+                        return scope;
+                    };
+
+                    scope.getData = function () {
+                        return scope.$modelValue;
+                    };
+
+                    scope.getElementChilds = function () {
+                        return angular.element(element[0].querySelector('[tree-dnd-nodes]'));
+                    };
+                }
             }
         };
     }